moved convenience and related tests to sandbox
