+++ selection/randomized/sandbox/convenience.py
