+++ selection/randomized/tests/sandbox/test_intervals.py
