+++ selection/randomized/tests/test_selective_MLE_high.py
@@ -2,12 +2,9 @@ import numpy as np
 import nose.tools as nt
 import rpy2.robjects as rpy
 from rpy2.robjects import numpy2ri
-#rpy.r('library(selectiveInference)')
 
-import selection.randomized.lasso as L; reload(L)
-from selection.randomized.lasso import lasso
-from selection.tests.instance import gaussian_instance
-import matplotlib.pyplot as plt
+from ..lasso import lasso, full_targets, selected_targets
+from ...tests.instance import gaussian_instance
 
 def test_full_targets(n=2000, p=200, signal_fac=0.5, s=5, sigma=3, rho=0.4, randomizer_scale=0.25, full_dispersion=True):
     """
@@ -15,42 +12,54 @@ def test_full_targets(n=2000, p=200, signal_fac=0.5, s=5, sigma=3, rho=0.4, rand
     """
 
     inst, const = gaussian_instance, lasso.gaussian
-    signal = np.sqrt(signal_fac * 2 * np.log(p))
-    X, Y, beta = inst(n=n,
-                      p=p, 
-                      signal=signal, 
-                      s=s,
-                      equicorrelated=False, 
-                      rho=rho, 
-                      sigma=sigma, 
-                      random_signs=True)[:3]
+    while True:
+        signal = np.sqrt(signal_fac * 2 * np.log(p))
+        X, Y, beta = inst(n=n,
+                          p=p, 
+                          signal=signal, 
+                          s=s,
+                          equicorrelated=False, 
+                          rho=rho, 
+                          sigma=sigma, 
+                          random_signs=True)[:3]
 
-    idx = np.arange(p)
-    sigmaX = rho ** np.abs(np.subtract.outer(idx, idx))
-    print("snr", beta.T.dot(sigmaX).dot(beta)/((sigma**2.)* n))
+        idx = np.arange(p)
+        sigmaX = rho ** np.abs(np.subtract.outer(idx, idx))
+        print("snr", beta.T.dot(sigmaX).dot(beta)/((sigma**2.)* n))
 
-    n, p = X.shape
+        n, p = X.shape
 
-    sigma_ = np.std(Y)
-    W = np.ones(X.shape[1]) * np.sqrt(2 * np.log(p)) * sigma_
+        sigma_ = np.std(Y)
+        W = np.ones(X.shape[1]) * np.sqrt(2 * np.log(p)) * sigma_
 
-    conv = const(X, 
-                 Y, 
-                 W, 
-                 randomizer_scale=randomizer_scale * sigma_)
-    
-    signs = conv.fit()
-    nonzero = signs != 0
-    print("dimensions", n, p, nonzero.sum())
+        conv = const(X, 
+                     Y, 
+                     W, 
+                     randomizer_scale=randomizer_scale * sigma_)
 
-    dispersion = None
-    if full_dispersion:
-        dispersion = np.linalg.norm(Y - X.dot(np.linalg.pinv(X).dot(Y)))**2 / (n - p)
+        signs = conv.fit()
+        nonzero = signs != 0
+        print("dimensions", n, p, nonzero.sum())
 
-    estimate, _, _, pval, intervals, _ = conv.selective_MLE(target="full", dispersion=dispersion)
+        if nonzero.sum() > 0:
+            dispersion = None
+            if full_dispersion:
+                dispersion = np.linalg.norm(Y - X.dot(np.linalg.pinv(X).dot(Y)))**2 / (n - p)
 
-    coverage = (beta[nonzero] > intervals[:,0]) * (beta[nonzero] < intervals[:,1])
-    return pval[beta[nonzero] == 0], pval[beta[nonzero] != 0], coverage, intervals
+            (observed_target, 
+             cov_target, 
+             cov_target_score, 
+             alternatives) = full_targets(conv.loglike, 
+                                          conv._W, 
+                                          nonzero, dispersion=dispersion)
+
+            estimate, _, _, pval, intervals, _ = conv.selective_MLE(observed_target, 
+                                                                    cov_target, 
+                                                                    cov_target_score, 
+                                                                    alternatives)
+
+            coverage = (beta[nonzero] > intervals[:,0]) * (beta[nonzero] < intervals[:,1])
+            return pval[beta[nonzero] == 0], pval[beta[nonzero] != 0], coverage, intervals
 
 def test_selected_targets(n=2000, p=200, signal_fac=1.5, s=5, sigma=3, rho=0.4, randomizer_scale=1, full_dispersion=True):
     """
@@ -59,41 +68,55 @@ def test_selected_targets(n=2000, p=200, signal_fac=1.5, s=5, sigma=3, rho=0.4, 
 
     inst, const = gaussian_instance, lasso.gaussian
     signal = np.sqrt(signal_fac * 2 * np.log(p))
-    X, Y, beta = inst(n=n,
-                      p=p, 
-                      signal=signal, 
-                      s=s, 
-                      equicorrelated=False, 
-                      rho=rho, 
-                      sigma=sigma, 
-                      random_signs=True)[:3]
 
-    n, p = X.shape
+    while True:
+        X, Y, beta = inst(n=n,
+                          p=p, 
+                          signal=signal, 
+                          s=s, 
+                          equicorrelated=False, 
+                          rho=rho, 
+                          sigma=sigma, 
+                          random_signs=True)[:3]
 
-    sigma_ = np.std(Y)
-    W = np.ones(X.shape[1]) * np.sqrt(2 * np.log(p)) * sigma_
+        n, p = X.shape
 
-    conv = const(X, 
-                 Y, 
-                 W, 
-                 randomizer_scale=randomizer_scale * sigma_)
-    
-    signs = conv.fit()
-    nonzero = signs != 0
+        sigma_ = np.std(Y)
+        W = np.ones(X.shape[1]) * np.sqrt(2 * np.log(p)) * sigma_
 
-    dispersion = None
-    if full_dispersion:
-        dispersion = np.linalg.norm(Y - X.dot(np.linalg.pinv(X).dot(Y)))**2 / (n - p)
+        conv = const(X, 
+                     Y, 
+                     W, 
+                     randomizer_scale=randomizer_scale * sigma_)
 
-    estimate, _, _, pval, intervals, _ = conv.selective_MLE(target="selected", dispersion=dispersion)
+        signs = conv.fit()
+        nonzero = signs != 0
 
-    beta_target = np.linalg.pinv(X[:,nonzero]).dot(X.dot(beta))
+        if nonzero.sum() > 0:
+            dispersion = None
+            if full_dispersion:
+                dispersion = np.linalg.norm(Y - X.dot(np.linalg.pinv(X).dot(Y)))**2 / (n - p)
 
-    coverage = (beta_target > intervals[:,0]) * (beta_target < intervals[:,1])
-    return pval[beta_target == 0], pval[beta_target != 0], coverage
+            (observed_target, 
+             cov_target, 
+             cov_target_score, 
+             alternatives) = selected_targets(conv.loglike, 
+                                              conv._W, 
+                                              nonzero, dispersion=dispersion)
+
+            estimate, _, _, pval, intervals, _ = conv.selective_MLE(observed_target, 
+                                                                    cov_target, 
+                                                                    cov_target_score, 
+                                                                    alternatives)
+
+            beta_target = np.linalg.pinv(X[:,nonzero]).dot(X.dot(beta))
+
+            coverage = (beta_target > intervals[:,0]) * (beta_target < intervals[:,1])
+            return pval[beta_target == 0], pval[beta_target != 0], coverage
 
 def main(nsim=500, full=True):
 
+    import matplotlib.pyplot as plt
     P0, PA, cover, length_int = [], [], [], []
     from statsmodels.distributions import ECDF
 
@@ -125,8 +148,6 @@ def main(nsim=500, full=True):
             if len(PA) > 0:
                 plt.plot(U, ECDF(PA)(U), 'r')
             plt.plot([0, 1], [0, 1], 'k--')
-            plt.savefig("/Users/snigdhapanigrahi/Desktop/plot.pdf")
     plt.show()
 
-main()
